refactor: split admin pages
