+++ src/App.tsx
@@ -43,6 +43,9 @@ const CompetitiveCharts = lazyWithRetry(() => import("./pages/CompetitiveCharts"
 const TouringSystem = lazyWithRetry(() => import("./pages/TouringSystem"));
 const Travel = lazyWithRetry(() => import("./pages/Travel"));
 const Admin = lazyWithRetry(() => import("./pages/Admin"));
+const AdminExperienceRewards = lazyWithRetry(() => import("./pages/admin/ExperienceRewards"));
+const AdminUniversities = lazyWithRetry(() => import("./pages/admin/Universities"));
+const AdminSkillBooks = lazyWithRetry(() => import("./pages/admin/SkillBooks"));
 const WorldEnvironment = lazyWithRetry(() => import("./pages/WorldEnvironment"));
 const SongManager = lazyWithRetry(() => import("./pages/SongManager"));
 const InventoryManager = lazyWithRetry(() => import("./pages/InventoryManager"));
@@ -113,6 +116,9 @@ function App() {
                     <Route path="charts-competitive" element={<CompetitiveCharts />} />
                     <Route path="tours-system" element={<TouringSystem />} />
                     <Route path="admin" element={<Admin />} />
+                    <Route path="admin/experience-rewards" element={<AdminExperienceRewards />} />
+                    <Route path="admin/universities" element={<AdminUniversities />} />
+                    <Route path="admin/skill-books" element={<AdminSkillBooks />} />
                     <Route path="world" element={<WorldEnvironment />} />
                     <Route path="songs" element={<SongManager />} />
                     <Route path="inventory" element={<InventoryManager />} />
