+++ src/pages/admin/universities.helpers.ts
@@ -1,0 +1,74 @@
+import { z } from "zod";
+
+import type { Database } from "@/integrations/supabase/types";
+
+export const UNIVERSITY_PAGE_SIZE = 10;
+
+export const sortColumnOptions = [
+  { value: "city", label: "City" },
+  { value: "name", label: "Name" },
+  { value: "prestige", label: "Prestige" },
+  { value: "quality_of_learning", label: "Quality of Learning" },
+  { value: "course_cost", label: "Course Cost" },
+] as const;
+
+export type SortColumn = (typeof sortColumnOptions)[number]["value"];
+export type SortDirection = "asc" | "desc";
+
+export const sortDirectionLabels: Record<SortDirection, string> = {
+  asc: "Ascending",
+  desc: "Descending",
+};
+
+export const universitySchema = z.object({
+  name: z.string().min(1, "Name is required"),
+  city: z.string().min(1, "City is required"),
+  prestige: z
+    .coerce
+    .number({ invalid_type_error: "Prestige must be a number" })
+    .min(0, "Prestige must be at least 0")
+    .max(100, "Prestige cannot exceed 100"),
+  qualityOfLearning: z
+    .coerce
+    .number({ invalid_type_error: "Quality must be a number" })
+    .min(0, "Quality must be at least 0")
+    .max(100, "Quality cannot exceed 100"),
+  courseCost: z
+    .coerce
+    .number({ invalid_type_error: "Course cost must be a number" })
+    .min(0, "Course cost cannot be negative"),
+});
+
+export type UniversityFormValues = z.infer<typeof universitySchema>;
+
+export type UniversitiesTable = Database["public"]["Tables"] extends { universities: infer T }
+  ? T
+  : {
+      Row: {
+        id: string;
+        name: string;
+        city: string;
+        prestige: number | null;
+        quality_of_learning: number | null;
+        course_cost: number | null;
+        created_at: string | null;
+      };
+      Insert: {
+        name: string;
+        city: string;
+        prestige?: number | null;
+        quality_of_learning?: number | null;
+        course_cost?: number | null;
+      };
+      Update: {
+        name?: string;
+        city?: string;
+        prestige?: number | null;
+        quality_of_learning?: number | null;
+        course_cost?: number | null;
+      };
+    };
+
+export type UniversityRow = UniversitiesTable extends { Row: infer R } ? R : never;
+export type UniversityInsert = UniversitiesTable extends { Insert: infer I } ? I : never;
+export type UniversityUpdate = UniversitiesTable extends { Update: infer U } ? U : never;
